# openapi.yaml
openapi: 3.0.3
info:
  title: Needle Eye API
  version: 0.1.0
  description: |
    Express + TypeScript backend for the Needle Eye boutique/tailoring ERP.
    Stateless REST: every request carries its own bearer token, verified
    independently -- nothing about a client's session is stored server-side
    between requests. `needleye-web` never talks to Supabase directly; every
    auth operation goes through the `/auth/*` endpoints documented here.

    See the repo README for the layered architecture (Controller ->
    Validation -> Service -> Repository -> Mapper) and the Mermaid flow-map
    diagrams -- this document covers the *what* (request/response contracts),
    the README's Flow Map section covers the *how* (which function calls
    which).
  contact:
    name: Needle Eye API
servers:
  - url: http://localhost:4000/api/v1
    description: Local dev
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List every staff account
      description: Owner/Manager only (`users:manage`).
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: { $ref: "#/components/schemas/User" }
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Users]
      summary: Create a staff account directly
      description: |
        Owner/Manager only. No email invite -- the account is created with a
        real password immediately (generated as "readable name + random
        suffix", e.g. `Aditya-9kQ2Xf7q`), returned once in this response.
        The Owner/Manager is responsible for communicating it to the new
        hire. The `handle_new_user()` DB trigger reads `full_name`/`role`
        from the created user's metadata and creates the matching `profiles`
        row automatically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, fullName, role]
              properties:
                email: { type: string, format: email }
                fullName: { type: string, minLength: 1 }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId: { type: string, format: uuid }
                  password: { type: string, description: "Shown only in this response -- never retrievable again." }
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/{id}/generate-password:
    post:
      tags: [Users]
      summary: Regenerate a staff member's password
      description: |
        Owner/Manager only. Always allowed for `designer`/`master_tailor`
        accounts. For `owner_manager`/`accountant` accounts, only allowed
        until that account's first real login (`profiles.last_login_at`) --
        after that they're expected to manage their own password via
        `POST /auth/password-reset-request` instead. The new password is
        returned once, same as account creation.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "201":
          description: Password regenerated
          content:
            application/json:
              schema:
                type: object
                properties: { password: { type: string } }
        "403":
          description: "This owner_manager/accountant account has already completed its first login"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{id}/qr-token:
    post:
      tags: [Users]
      summary: Generate (or regenerate) a Master Tailor's QR login code
      description: |
        Owner/Manager only, and only for `master_tailor` accounts (400 for
        any other role). Only the token's SHA-256 hash is stored
        (`qr_login_tokens`, a table with no anon/authenticated grant at
        all -- unreachable except through this API's service-role client).
        The raw token/URL is returned once; regenerating immediately
        invalidates whatever QR code was printed/displayed before.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "201":
          description: QR login token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  loginUrl: { type: string, example: "http://localhost:3000/qr-login?token=..." }
        "400":
          description: The target account is not a master_tailor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{id}:
    patch:
      tags: [Users]
      summary: Update a staff member's name, role, or active flag
      description: Owner/Manager only. All fields optional; at least one is required.
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName: { type: string, minLength: 1 }
                role: { $ref: "#/components/schemas/Role" }
                active: { type: boolean }
      responses:
        "204": { description: Updated }
        "400":
          description: No fields to update
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/{id}/deactivate:
    post:
      tags: [Users]
      summary: Deactivate a staff member
      description: |
        Owner/Manager only. Sets `profiles.active = false` AND bans the
        account at the Supabase Auth layer (`ban_duration`), so the block is
        enforced at sign-in, not just app-level. Also clears any QR login
        token immediately -- a deactivated account's QR must stop working
        right away, not linger. Rejects self-deactivation.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204": { description: Deactivated }
        "400":
          description: You cannot deactivate your own account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase-issued access token, obtained from `POST /auth/login` (or `/refresh`, `/exchange-code`). Verified independently on every request -- see `common/middleware/auth.middleware.ts`.
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    Forbidden:
      description: Authenticated, but the caller's role/scope doesn't allow this
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found (or outside the caller's row scope)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      description: The one shape every error response takes, produced by common/middleware/error.middleware.ts regardless of which layer threw.
      properties:
        error: { type: string }
        code: { type: string, example: FORBIDDEN }
        details: {}
    Role:
      type: string
      enum: [owner_manager, designer, master_tailor, accountant]
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time, nullable: true, description: "Gates whether generate-password stays available for owner_manager/accountant accounts." }
        hasQrLogin: { type: boolean, description: "Whether a QR login token exists. Never the token/hash itself." }
...